pyproject.toml
[project]
name = "anylist-cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=46.0.3",
    "google>=3.0.0",
    "httpx>=0.28.1",
    "pre-commit>=4.3.0",
    "protobuf>=6.33.1",
    "python-dotenv>=1.2.1",
]
